Minor edits

+++ src/problem5/contracts/Balance.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.13;
 
 interface IERC20 {
-    function balanceOf(address account) external view returns (uint);
+    function balanceOf(address account) external view returns (uint256);
 }
 
 contract Balance {
